ConnectionInfo
NodeInput and NodeOutput can be deleted. They can then be reconstructed by creating them again. Despite the objects providing information about the input and output, they aren't really needed to determine a connection.

To avoid the problem of having to determine a connection based on two classes, a ConnectionInfo class was made to hold the input, output and corresponding indices. This helps solve problems when deleting a port dynamically (especially for the cognix editor).

(Truth is, NodeInput and NodeOutput could handle the indices internally, but that was a bigger undertaking)